+++ frontend/sfem_ForwardDeclarations.hpp
@@ -1,0 +1,15 @@
+#ifndef SFEM_FORWARD_DECLARATIONS_HPP
+#define SFEM_FORWARD_DECLARATIONS_HPP
+
+namespace sfem {
+    class Function;
+    class Mesh;
+    class FunctionSpace;
+    class Op;
+    class CRSGraph;
+
+    class DirichletConditions;
+    class NeumannConditions;
+}
+
+#endif //SFEM_FORWARD_DECLARATIONS_HPP
